Got a Web front end running; some test cleanup.
OpenLocalGitRepo now traverses upwards in a directory tree to find the first
parent with a ".git" directory under it. This is useful for opening a repo
based on an absolute path to a file, which is what the ugly Web front end
currently takes as input.

Moved tempGitRepo stuff to new test_util package.